+++ CHANGES.rst
@@ -16,7 +16,7 @@ in progress
   telling PR- and nightly releases apart from GA releases when running
   ``mqttwarn --version``.
 - Improve Apprise integration by propagating the mqttwarn data dictionary into
-  the Apprise plugin template arguments.
+  the Apprise plugin template arguments. Thanks, @zoic21.
 
 
 2022-11-21 0.31.0