+++ CHANGES.rst
@@ -7,6 +7,7 @@ in progress
 ===========
 
 - Pushsafer: Fix to prevent submitting empty parameters to upstream API.
+- Pushsafer: Modernize configuration layout for target addresses.
 
 
 2023-02-13 0.32.0