+++ CHANGES.rst
@@ -6,9 +6,15 @@ mqttwarn changelog
 in progress
 ===========
 
+
+2023-04-11 0.33.0
+=================
+
 - Pushsafer: Fix to prevent submitting empty parameters to upstream API.
 - Pushsafer: Modernize configuration layout for target addresses.
 - Pushsafer: Add parameters for "Confirm", "Answer Options", and "Force Answer".
+
+Thanks, @appzer.
 
 
 2023-02-13 0.32.0